+++ server/src/datastore.js
@@ -18,3 +18,14 @@ export const getComponents = (...a) => impl.getComponents(...a);
 export const getActivePlayerShip = (...a) => impl.getActivePlayerShip(...a);
 export const getMap = (...a) => impl.getMap(...a);
 export const getLevel = (...a) => impl.getLevel(...a);
+// Authentication (DECISIONS §11)
+export const getPlayerPublic = (...a) => impl.getPlayerPublic(...a);
+export const setUsername = (...a) => impl.setUsername(...a);
+export const findPlayerForLogin = (...a) => impl.findPlayerForLogin(...a);
+export const emailInUse = (...a) => impl.emailInUse(...a);
+export const registerAccount = (...a) => impl.registerAccount(...a);
+export const setVerifyToken = (...a) => impl.setVerifyToken(...a);
+export const verifyEmailToken = (...a) => impl.verifyEmailToken(...a);
+export const createSession = (...a) => impl.createSession(...a);
+export const getSessionPlayer = (...a) => impl.getSessionPlayer(...a);
+export const deleteSession = (...a) => impl.deleteSession(...a);